requirements-dev.txt
-r requirements.txt
black==22.10.0
flake8==5.0.4
isort==5.12.0
mock~=4.0.3
pre-commit~=2.20.0
pytest-asyncio~=0.20.2
pytest-django~=4.5.2
pytest-dotenv~=0.5.0
pytest-httpx~=0.21
